requirements.txt
python-telegram-bot
feedparser
requests
beautifulsoup4
